Mer DRY navigation bar
